Added check to automatically change color of banner if the WordPress category changes.  Ready for testing.

+++ lib.rss.php
@@ -75,6 +75,7 @@
     // Takes no arguements
     function rmData()
     {
-        unlink('emergency');
+        if (file_exists('emergency'))
+            unlink('emergency');
     }
 ?>
